# Conflicts:
#	.gitignore
#	src/main/java/kyung/kung_backend/domain/expert/dto/ExpertDetailResponse.java
#	src/main/java/kyung/kung_backend/domain/expert/service/ExpertService.java
#	src/main/java/kyung/kung_backend/domain/servicepost/repository/ExpertServiceRepository.java
#	src/main/java/kyung/kung_backend/domain/servicepost/service/ExpertServicePostService.java
#	src/main/java/kyung/kung_backend/global/jwt/JwtProvider.java
#	src/main/java/kyung/kung_backend/global/security/SecurityConfig.java

+++ src/main/java/kyung/kung_backend/domain/favorite/dto/FavoriteExpertResponse.java
@@ -10,7 +10,7 @@ public class FavoriteExpertResponse {
 
     private Long expertProfileId;
     private String displayName;
-    private Long careerYears;
+    private Double careerYears;
     private String mainCategoryName;
     private boolean favorite;
